+++ packages/fx-core/src/plugins/resource/frontend/dotnet/ops/deploy.ts
@@ -21,12 +21,12 @@ import {
   DotnetPluginInfo as PluginInfo,
   RegularExpr,
 } from "../constants";
-import { forEachFileAndDir } from "../../../function/utils/dir-walk";
 import { Logger } from "../../utils/logger";
 import { Messages } from "../resources/messages";
 import { ProgressHelper } from "../../utils/progress-helper";
 import { WebappDeployProgress as DeployProgress } from "../resources/steps";
 import { Utils } from "../../utils";
+import { forEachFileAndDir } from "../../../../../component/utils/fileOperation";
 
 export async function getFrameworkVersion(projectFilePath: string): Promise<string> {
   const content = await fs.readFile(projectFilePath, "utf8");
