--- packages/fx-core/src/plugins/resource/function/language-strategy.ts
@@ -1,79 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-import * as path from "path";
-
-import { Commands, CommonConstants, FunctionPluginPathInfo } from "./constants";
-import { FunctionLanguage } from "./enums";
-
-export interface FunctionLanguageStrategy {
-  /* For scaffolding. */
-  getFunctionEntryFileOrFolderName: (entryName: string) => string;
-
-  /* For deployment. */
-  skipFuncExtensionInstall: boolean;
-  hasUpdatedContentFilter?: (itemPath: string) => boolean;
-  buildCommands: {
-    command: string;
-    relativePath: string;
-  }[];
-  deployFolderRelativePath: string;
-}
-
-const NodeJSCommonStrategy: FunctionLanguageStrategy = {
-  getFunctionEntryFileOrFolderName: (entryName: string) => entryName,
-  skipFuncExtensionInstall: false,
-  /* We skip scanning node_modules folder for node because it has too many small files.
-   * Its consistency can be guaranteed by `npm install`.
-   */
-  hasUpdatedContentFilter: (itemPath: string) =>
-    path.basename(itemPath) !== FunctionPluginPathInfo.npmPackageFolderName,
-  buildCommands: [],
-  deployFolderRelativePath: CommonConstants.emptyString,
-};
-
-const JavaScriptLanguageStrategy: FunctionLanguageStrategy = {
-  ...NodeJSCommonStrategy,
-  buildCommands: [
-    {
-      command: Commands.npmInstallProd,
-      relativePath: CommonConstants.emptyString,
-    },
-  ],
-};
-
-const TypeScriptLanguageStrategy: FunctionLanguageStrategy = {
-  ...NodeJSCommonStrategy,
-  buildCommands: [
-    {
-      command: Commands.npmInstall,
-      relativePath: CommonConstants.emptyString,
-    },
-    {
-      command: Commands.npmBuild,
-      relativePath: CommonConstants.emptyString,
-    },
-  ],
-};
-
-// const CSharpLanguageStrategy: FunctionLanguageStrategy = {
-//     getFunctionEntryFileOrFolderName: (entryName: string) => `${entryName}.cs`,
-//     skipFuncExtensionInstall: true,
-//     buildCommands: [{
-//         command: Commands.dotnetPublish,
-//         relativePath: CommonConstants.emptyString
-//     }],
-//     deployFolderRelativePath: path.join("bin", "Release", "netcoreapp3.1", "publish")
-// };
-
-export class LanguageStrategyFactory {
-  public static getStrategy(language: FunctionLanguage): FunctionLanguageStrategy {
-    switch (language) {
-      case FunctionLanguage.JavaScript:
-        return JavaScriptLanguageStrategy;
-      case FunctionLanguage.TypeScript:
-        return TypeScriptLanguageStrategy;
-      // case FunctionLanguage.CSharp:
-      //     return CSharpLanguageStrategy;
-    }
-  }
-}
