--- packages/fx-core/src/plugins/resource/function/utils/dir-walk.ts
@@ -1,23 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-import * as fs from "fs-extra";
-import klaw from "klaw";
-
-export async function forEachFileAndDir(
-  root: string,
-  callback: (itemPath: string, stats: fs.Stats) => boolean | void,
-  filter?: (itemPath: string) => boolean
-): Promise<void> {
-  await new Promise((resolve, reject) => {
-    const stream: klaw.Walker = klaw(root, { filter: filter });
-    stream
-      .on("data", (item) => {
-        if (callback(item.path, item.stats)) {
-          stream.emit("close");
-        }
-      })
-      .on("end", () => resolve({}))
-      .on("error", (err) => reject(err))
-      .on("close", () => resolve({}));
-  });
-}
