+++ packages/fx-core/src/plugins/resource/index.ts
@@ -4,6 +4,5 @@
 
 export * from "./apim";
 export * from "./frontend";
-export * from "./function";
 export * from "./identity";
 export * from "./localdebug";
